+++ src/main/java/org/truffleruby/parser/YARPTranslator.java
@@ -175,6 +175,9 @@ public class YARPTranslator extends YARPBaseTranslator {
     public static final RescueNode[] EMPTY_RESCUE_NODE_ARRAY = new RescueNode[0];
 
     public Deque<Integer> frameOnStackMarkerSlotStack = new ArrayDeque<>();
+
+    protected RubyDeferredWarnings rubyWarnings;
+
     /** whether a while-loop body is translated; needed to check correctness of operators like break/next/etc */
     private boolean translatingWhile = false;
     /** whether a for-loop body is translated; needed to enforce variables in the for-loop body to be declared outside
@@ -198,8 +201,9 @@ public class YARPTranslator extends YARPBaseTranslator {
     /** all the encountered BEGIN {} blocks; they will be added finally at the beginning of the program AST */
     private final ArrayList<RubyNode> beginBlocks = new ArrayList<>();
 
-    public YARPTranslator(TranslatorEnvironment environment) {
+    public YARPTranslator(TranslatorEnvironment environment, RubyDeferredWarnings rubyWarnings) {
         super(environment);
+        this.rubyWarnings = rubyWarnings;
     }
 
     public RubyNode[] getBeginBlocks() {
@@ -537,7 +541,9 @@ public class YARPTranslator extends YARPBaseTranslator {
 
         final YARPBlockNodeTranslator methodCompiler = new YARPBlockNodeTranslator(
                 newEnvironment,
-                arity);
+                arity,
+                parameters.block != null,
+                rubyWarnings);
 
         methodCompiler.frameOnStackMarkerSlotStack = frameOnStackMarkerSlotStack;
 
@@ -1548,7 +1554,8 @@ public class YARPTranslator extends YARPBaseTranslator {
 
         final var defNodeTranslator = new YARPDefNodeTranslator(
                 language,
-                newEnvironment);
+                newEnvironment,
+                rubyWarnings);
         var callTargetSupplier = defNodeTranslator.buildMethodNodeCompiler(node, parameters, arity);
 
         final boolean isDefSingleton = singletonClassNode != null;
@@ -3562,7 +3569,7 @@ public class YARPTranslator extends YARPBaseTranslator {
             newEnvironment.declareVar(name);
         }
 
-        final YARPTranslator moduleTranslator = new YARPTranslator(newEnvironment);
+        final YARPTranslator moduleTranslator = new YARPTranslator(newEnvironment, rubyWarnings);
         final ModuleBodyDefinition definition = moduleTranslator.compileClassNode(moduleNode, bodyNode);
         return new RunModuleDefinitionNode(definition, defineOrGetNode);
     }